@@ -10,19 +10,19 @@
   },
   "devDependencies": {
     "@iconify/svelte": "^3.1.6",
-    "@sveltejs/vite-plugin-svelte": "^3.0.1",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/typography": "^0.5.10",
     "autoprefixer": "^10.4.16",
     "postcss": "^8.4.33",
     "postcss-pxtorem": "^6.0.0",
     "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
+    "prettier-plugin-svelte": "^3.2.6",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "rfs": "^10.0.0",
-    "svelte": "^4.2.8",
-    "svelte-routing": "^2.12.0",
+    "svelte": "^5.0.0",
+    "svelte5-router": "^3.0.2",
     "tailwindcss": "^3.4.1",
-    "vite": "^5.0.11"
+    "vite": "^5.4.4"
   },
   "dependencies": {
     "plyr": "^3.7.8",
