@@ -1,12 +1,18 @@
 <script>
   import { onMount } from "svelte";
-  import { navigate, useLocation } from "svelte-routing";
+  import { navigate, useLocation } from "svelte5-router";
   import { currentUser } from "../lib/pocketbase";
   import Icon from "@iconify/svelte";
   import Title from "../components/Title.svelte";
   import { t } from "../lib/i18n";
 
-  export let page = "page";
+  /**
+   * @typedef {Object} Props
+   * @property {string} [page]
+   */
+
+  /** @type {Props} */
+  let { page = "page" } = $props();
   const location = useLocation();
 
   onMount(() => {
@@ -35,7 +41,7 @@
         > does not exist.
       </p>
       <button
-        on:click={() => navigate("/")}
+        onclick={() => navigate("/")}
         class="flex items-center justify-center gap-2 rounded-md bg-main px-4 py-2 transition hover:bg-main/80 2xs:w-full 2xs:px-0"
       >
         <Icon class="flex-shrink-0" icon="ph:arrow-left" />
